requirements.txt
flask>=2.0.3
line-bot-sdk>=2.1.0
qrcode>=7.3.1
gunicorn>=20.1.0
